refactor(calculator): long[] 타입을 double[]으로 변경

+++ src/main/java/calculator/Application.java
@@ -8,7 +8,7 @@ public class Application {
         System.out.println("덧셈할 문자열을 입력해 주세요.");
 
         String userInput = Console.readLine();
-        long result = StringCalculator.calculate(userInput);
+        double result = StringCalculator.calculate(userInput);
 
         System.out.println("결과 : " + result);
     }
